--- aspnet/web-api/overview/advanced/calling-a-web-api-from-a-net-client/sample/server/Controllers/ProductsController.cs
@@ -1,60 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net;
-using System.Net.Http;
-using System.Web.Http;
-using ProductStore.Models;
-
-namespace ProductStore.Controllers
-{
-    public class ProductsController : ApiController
-    {
-        static readonly IProductRepository repository = new ProductRepository();
-
-        public IEnumerable<Product> GetAllProducts()
-        {
-            return repository.GetAll();
-        }
-
-        public Product GetProduct(int id)
-        {
-            Product item = repository.Get(id);
-            if (item == null)
-            {
-                throw new HttpResponseException(HttpStatusCode.NotFound);
-            }
-            return item;
-        }
-
-        public IEnumerable<Product> GetProductsByCategory(string category)
-        {
-            return repository.GetAll().Where(
-                p => string.Equals(p.Category, category, StringComparison.OrdinalIgnoreCase));
-        }
-
-        public HttpResponseMessage PostProduct(Product item)
-        {
-            item = repository.Add(item);
-            var response = Request.CreateResponse<Product>(HttpStatusCode.Created, item);
-
-            string uri = Url.Link("DefaultApi", new { id = item.Id });
-            response.Headers.Location = new Uri(uri);
-            return response;
-        }
-
-        public void PutProduct(int id, Product product)
-        {
-            product.Id = id;
-            if (!repository.Update(product))
-            {
-                throw new HttpResponseException(HttpStatusCode.NotFound);
-            }
-        }
-
-        public void DeleteProduct(int id)
-        {
-            repository.Remove(id);
-        }
-    }
-}
