--- aspnet/web-api/overview/advanced/calling-a-web-api-from-a-net-client/sample/server/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="ProductsApp.WebApiApplication" Language="C#" %>
