--- aspnet/web-api/overview/advanced/calling-a-web-api-from-a-net-client/sample/server/Models/IProductRepository.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ProductStore.Models
-{
-    interface IProductRepository
-    {
-        IEnumerable<Product> GetAll();
-        Product Get(int id);
-        Product Add(Product item);
-        void Remove(int id);
-        bool Update(Product item);
-    }
-}
