--- aspnet/web-api/overview/advanced/calling-a-web-api-from-a-net-client/sample/server/Models/Product.cs
@@ -1,10 +1,0 @@
-﻿namespace ProductStore.Models
-{
-    public class Product
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-        public string Category { get; set; }
-        public decimal Price { get; set; }
-    }
-}
